Add files via upload
A Spotify clone is a web or mobile application designed to replicate the features and functionality of the popular music streaming platform Spotify. This type of project is often developed for learning purposes or as part of a portfolio to showcase skills in modern technologies such as web development, app development, and backend systems.